axis markers

+++ include/bootleg/game.hpp
@@ -103,13 +103,13 @@ struct Rotation {
         return Rotation(angle, 0, 1, 0);
     }
 };
-Vector2 measure_codepoint_3d(int codepoint, Font font, int font_size);
-Vector2 measure_text_3d(const char* txt, Font font, int font_size, int spacing);
-void draw_text_3d(const char* txt, Font font, Vector3 pos, int font_size,
+Vector2 measure_codepoint_3d(int codepoint, Font font, float font_size);
+Vector2 measure_text_3d(const char* txt, Font font, float font_size, int spacing);
+void draw_text_3d(const char* txt, Font font, Vector3 pos, float font_size,
     int spacing, const Color& color,
     bool backface = false, const Rotation& rotation = Rotation::none());
 Vector2 draw_codepoint_3d(int codepoint, Font font, const Vector3& pos,
-    int font_size, const Color& color,
+    float font_size, const Color& color,
     bool backface = false, const Rotation& rotation = Rotation::none());
 
 class Game;
